DDF-718 - Fixed issue with file separators in the  ApplicationFileInstaller.
There was an issue with the file upload feature on the add application modal in the admin ui.  Seems that running this on windows the string for the URI wasn't getting the correct file separators.  Before it would work fine on linux based machines and macs but on windows, you get both / and \ in the same string which makes the URI framework barf.

Also fixed issue with file types in the html input element.  should
support what the backend supports...kar and jar.