DDF-3584 Adds `--bail` flag to `catalog-ui-search` package
Previously, any webpack errors would be ignored and the result would be
a wab that could be deployed but might not work properly at runtime. To
detect these failures at build-time, the `--bail` flag informs webpack
to exit with a non-zero status which results in the failure of the
overall build.