DDF-4357 Select the first item in menus when opened (#4075)
* DDF-4357 Select the first item in menus when opened

* DDF-4357 Move state selection or clearing logic from `render` to `componentDidUpdate`

Modifying state during render is dangerous in React; `componentDidUpdate` can safely contain state modifications. Note that `componentDidUpdate` is not called upon first render, but the `Menu` constructor currently contains logic to select the first item in the menu if there are any. This first item selection logic in the constructor did not accomplish the goal of selecting the first item in the Gazetteer search box because that search box actually renders first with no children while the suggestions are being retrieved from Solr, then rerenders with new children without selecting the first of the new children.

* DDF-4357 Factored out active choice logic and applied it to state initialization

* DDF-4357 Format

* DDF-4357 Simplify `chooseActive` logic and fix a potential bug

As it turns out, `active` should be cleared not only if the menu is rerendered with no children, but also when the actively selected menu item is not longer in the menu items. The logic was changed to reflect this and simultaneously to simplify.